+++ MBF/config/config.json
@@ -2,7 +2,7 @@
   "development": {
     "username": "root",
     "password": "root",
-    "port": 3306,
+    "port": 3309,
     "database": "exampledb",
     "host": "localhost",
     "dialect": "mysql"
