+++ MBF/package.json
@@ -15,7 +15,10 @@
     "dotenv": "^6.0.0",
     "express": "^4.16.3",
     "express-handlebars": "^3.0.0",
-    "mysql2": "^1.5.3",
+    "fs": "0.0.1-security",
+    "mysql": "^2.16.0",
+    "mysql2": "^1.6.1",
+    "path": "^0.12.7",
     "sequelize": "^4.38.0"
   },
   "devDependencies": {
